+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## Master
 
 #### Breaking
+- `RequestSerializer` no longer handles query parameters
 - `XMLNode` value getters have been updated to conform to `XMLTextNodeInitializable`.
 
 #### Enhancements
+- `FormEncodedRequestSerializer` now exposes query formatting options for body parameters
+- `QueryStringFormattingOptions` now encodes plus symbols by default
 - `XMLTextNodeInitializable` protocol has been added.
 
 #### Bug Fixes